chore: move templayed.js to centralized folder
